Move edit utils file
